Use regular integer

+++ database/migrations/2016_09_01_010508_create_bvams_table.php
@@ -27,7 +27,7 @@ class CreateBvamsTable extends Migration
 
             $table->timestamp('first_validated_at')->nullable()->index();
             $table->timestamp('last_validated_at')->nullable();
-            $table->tinyInteger('confirmations')->nullable()->unsigned();
+            $table->integer('confirmations')->nullable()->unsigned();
 
             $table->index('asset');
         });
